+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ### Added
 - 'About' page: Shows automatically when extension first installed, or via button in menu.
 - TODO: Asana: New trigger with user-defined section names for drag/drop event in board view.
+- Wrike: New triggers for all 'Completed' buttons and checkboxes on Wrike.com
+- Jira: New triggers for <user-defined-status> buttons on *.atlassian.net
 
 ## [0.1.1] - 2022-03-12
 ### Added